Rename variables, split up, use openfisca as default, remove unused
